added test added datamodel class
